Fix id type

+++ backend/src/main/java/application/controller/TaskController.java
@@ -31,12 +31,12 @@ public class TaskController {
     }
 
     @PutMapping("/employees/{id}")
-    Task setPreviousTask(@RequestBody Task previousTask, @PathVariable Integer id) {
+    Task setPreviousTask(@RequestBody Task previousTask, @PathVariable Long id) {
             return taskService.setPreviousTask(id, previousTask);
     }
 
     @PutMapping("/employees/{id}")
-    Task updateTask(@RequestBody Task newTask, @PathVariable Integer id) {
+    Task updateTask(@RequestBody Task newTask, @PathVariable Long id) {
         return taskService.updateTask(id, newTask);
     }
